requirements.txt
amqp==5.2.0
anyio==4.4.0
asgiref==3.8.1
billiard==4.2.0
boto3==1.35.5
botocore==1.35.5
celery==5.4.0
channels==4.1.0
channels-redis==4.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.1
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-storages==1.14.4
django-timezone-field==7.0
djangorestframework==3.15.2
ffmpeg-python==0.2.0
future==1.0.0
h11==0.14.0
httptools==0.6.1
idna==3.8
itsdangerous==2.2.0
jmespath==1.0.1
kombu==5.4.0
msgpack==1.0.8
prompt_toolkit==3.0.47
psycopg==3.2.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytubefix==6.13.0
PyYAML==6.0.2
redis==5.0.8
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.20.0
vine==5.1.0
watchfiles==0.23.0
wcwidth==0.2.13
websockets==13.0
